The Shane Clause
Since we’re already calculating the number of ladies per show, no that value is passed through to the loop for getting the data. If that number is greater than 20, we use the number, otherwise we default to only finding the first 20. This causes all shows with a reasonable number of WLW to load about .005 seconds faster on average. It doesn’t help The L Word, because Shane is promiscuous.