Listicles
* Bumping version to break CSS caching
* Tweak where we set the counter for Gutenberg being an idiot

The problem being faced is that nesting nesting blocks is incredibly 
complicated and messy. So instead of that, I went with a templated 
block. You add that to the post and it listicles.

The downside to this is you can only have ONE Gutenlisticle per page, 
but I'm hoping that's not an issue. If it is, I'll have to do something 
else...

The CSS for Gutenlisticles is in the Block for REASONS.